Split database module into two source files
